# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/property-categories/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a propertyCategory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PropertyCategories"
                ],
                "summary": "Create a propertyCategory",
                "parameters": [
                    {
                        "description": "Create a propertyCategory",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.CreatePropertyCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PropertyCategory response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/property-categories/get-by-filters": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get PropertyCategories",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PropertyCategories"
                ],
                "summary": "Get PropertyCategories",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PropertyCategory response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.PagedList-github_com_Aliizi83_sample-golang-project_src_api_dto_PropertyCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/property-categories/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a propertyCategory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PropertyCategories"
                ],
                "summary": "Get a propertyCategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "propertyCategory response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a propertyCategory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PropertyCategories"
                ],
                "summary": "Update a propertyCategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a propertyCategory",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.UpdatePropertyCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PropertyCategory response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a propertyCategory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PropertyCategories"
                ],
                "summary": "Delete a propertyCategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.CreatePropertyCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                },
                "dataType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_api_dto.UpdatePropertyCategoryRequest": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_api_helpers.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.PagedList-github_com_Aliizi83_sample-golang-project_src_api_dto_PropertyCategoryResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_api_dto.PropertyCategoryResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sorts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Aliizi83_sample-golang-project_src_domain_filters.Sort"
                    }
                }
            }
        },
        "github_com_Aliizi83_sample-golang-project_src_domain_filters.Sort": {
            "type": "object",
            "properties": {
                "columnId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}